made it so divide returns double instead of int

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -14,7 +14,8 @@ public class MyMathCalc {
     public int multiply(int a, int b){
         return a*b;
     }
-    public float divide(int a, int b){
+
+    public double divide(int a, int b){
         return a/b;
     }
 }
